Backwards for binary operations with scalars
