+++ src/autograd/mod.rs
@@ -1,3 +1,4 @@
 pub mod gradient_function;
 pub mod into;
+pub mod ops_backwards;
 pub mod backwards;
